[IMP] website, *: move utility functions from website to web_editor
