+++ addons/website/static/src/snippets/s_embed_code/options.js
@@ -4,7 +4,7 @@ import { useService } from "@web/core/utils/hooks";
 import options from '@web_editor/js/editor/snippets.options';
 import { _t } from "@web/core/l10n/translation";
 import { EditHeadBodyDialog } from "@website/components/edit_head_body_dialog/edit_head_body_dialog";
-import { cloneContentEls } from "@website/js/utils";
+import weUtils from "@web_editor/js/common/utils";
 
 import { Component, useState } from "@odoo/owl";
 
@@ -58,8 +58,8 @@ options.registry.EmbedCode = options.Class.extend({
                     // Removes scripts tags from the DOM as we don't want them
                     // to interfere during edition, but keeps them in a
                     // `<template>` that will be saved to the database.
-                    templateEl.content.replaceChildren(cloneContentEls(newValue, true));
-                    $container[0].replaceChildren(cloneContentEls(newValue));
+                    templateEl.content.replaceChildren(weUtils.cloneContentEls(newValue, true));
+                    $container[0].replaceChildren(weUtils.cloneContentEls(newValue));
                 }
             }, {
                 onClose: resolve,
