+++ addons/website/static/src/systray_items/website_switcher.js
@@ -4,7 +4,7 @@ import { registry } from "@web/core/registry";
 import { useService } from "@web/core/utils/hooks";
 import { Dropdown } from "@web/core/dropdown/dropdown";
 import { DropdownItem } from "@web/core/dropdown/dropdown_item";
-import wUtils from '@website/js/utils';
+import weUtils from "@web_editor/js/common/utils";
 import { Component } from "@odoo/owl";
 
 export class WebsiteSwitcherSystray extends Component {
@@ -26,7 +26,7 @@ export class WebsiteSwitcherSystray extends Component {
             id: website.id,
             domain: website.domain,
             callback: () => {
-                if (website.domain && !wUtils.isHTTPSorNakedDomainRedirection(website.domain, window.location.origin)) {
+                if (website.domain && !weUtils.isHTTPSorNakedDomainRedirection(website.domain, window.location.origin)) {
                     const { location: { pathname, search, hash } } = this.websiteService.contentWindow;
                     const path = pathname + search + hash;
                     window.location.href = `${encodeURI(website.domain)}/odoo/action-website.website_preview?path=${encodeURIComponent(path)}&website_id=${encodeURIComponent(website.id)}`;
